requirements.txt
opencv-python
tensorflow
pillow
imgaug